Modified file structure...added serializers file
